requirements.txt
boto3==1.42.35
botocore==1.42.35
faiss-cpu==1.13.2
jmespath==1.1.0
numpy==2.4.1
packaging==26.0
pypdf==6.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
